Started working on Simd like problem.
I like the way I was able to formulate this.  The formulation
of IsNumber loops over all the squares on the board and
just introduces a constraint that that square's number
is a valid solution.  I suspect this will make solving very,
very fast (about as fast as possible).  Hopefully all the
constraints can work that way (I think they can).

Speed is important because my next step will be generation
of all possible satisfied constraints and puzzle generation
should be much faster.