Functional puzzle generator
I seem to have created a puzzle generator that works.  I
went with an approach that sketches out not just the
board and the solution but also a "plan".  This plan allows
me to define the order of deductions.  This helps guide the
creation of constraints so that you don't end up with stupid
problems where there are constraints that define secondary
variables but those variables don't affect any other variables.

The plan is a hierarchical description of the solution order.  I
think it is rich enough to create sufficiently interesting or
complex problem.

What is current missing is the "trim" capability to get rid
of redundant constraints.  I'm not sure if this should be
done locally as we construct constraints or globally at
the end.  I have to think about which is more efficient
but also make sure it doesn't negatively impact the "fun"
factor of each puzzle.

Ultimately, what I expect to happen is that through some
random process we will synthesize a board, a solution
and a plan.  Then the puzzle builder will identify a set
of clues that are consistent and well specified subject
to those constraints.

For harder puzzles, I think it will be necessary/interesting
for secondary variables to have multiple solutions.  In other
words, the criteria for "trimming" the system should be
that the primary variable always has a unique solution but
that the secondary variables are not required to have a
unique solution.  The number of solutions for secondary
variables could be used as a measure of "hardness".

Finally, although we currently synthesize the constraints
to be consistent with a specifical a priori solution.  When
we trim, we don't need to maintain that unique solution,
only the unique solution on the primary variable.