Actually working
I fixed a bug in trim and added a random board
generator.  Plugging those together got the generator
synthesizing puzzles.  Even more interesting, the
current puzzle it spits out (which is deterministic
because of the random seed) is actually quite interesting.
I'm going to remove IsNumber though, it is too direct.