Created a new problem generator
I call this the Scuptor approach.  It generates all valid
constraints.  At that point, it confirms that a solution
exists for the specified solution.  Once that is established,
it randomizes and sorts (by priority) those constraints and
attempts to eliminate them one by one.  If eliminating
a constraint would break the "plan" (represented by the
DTree class) for dependencies between variables, then
the constraint cannot be eliminated.  If the constraint
doesn't break the plan, then we check to make sure a
unique solution exists if the constraint is eliminated.  If
so, the constraint is removed...otherwise we keep it.

This seems to work (even for purely secondary constraints!).
But the number of constraints ends up being pretty high.
I'm curious whether this could be a measure of difficulty?