Turned on debugging
The result is that exploring all combinations of constraints is very time
consuming.  One potential optimization would be to find a way to prune
the tree.  Another might be to find another way to collect the key
metric of maxc required to solve for any one variable.  Perhaps a smarter
(less brute force) algorithm is required here.  I need to think about it.