Refactored generators and it looks good.
There is some weird issue I'm having with for
comprehensions and option types.  I converted
the options to a List and it fixed the issue.  I think it
might be some kind of bug in IntelliJ, the presentation
compiler, the fast compiler or whatever zinc-y kind
thing they are doing to compile because it looks
perfectly valid to me.