Basic puzzle builder working
There is an issue here though that it seems to be leaving
redundant constraints in the mix.  I'm not sure if it is a
bug in the constraint implementation or the puzzle builder.
Other than that, it seems to be working as expected and
it is very simple (and I suspect quite efficient due to the
solver on the stack approach).