All tests pass, need more optimization
I've implemented a solver that uses a Stream which,
if I understand it correctly, implements some kind of
lazy evaluation.  My goal is to switch completely to
a Stream based approach and then I can use this to
optimize things by avoid a complete search in cases where
I only need to know if there are multiple solutions.