All tests passing again
I noted that LessThan and GreaterThan were the only
asymmetric SecondaryConstraints.  Then I started looking
into how the SetSolver worked to make sure there was
nothing strange with the way SecondaryConstraints were
solved.  I noticed that the code the check for consistency
of constraints didn't use the new "consistent" method
in Constraint.  I changed it to use that and then the issue
resolved itself (LessThan and GreaterThan are back in
the generators).  However, it did create a new performance
issue that I need to look into.  The current performance is
acceptable, but I'd like to see if there are any obvious
ways to speed it up.