Performance issue identified
I was doing a really inefficient way of checking consistency
of constraints that involved an unnecessary loop.  I removed
it and used something much smarter, simpler and faster
and all the tests continue to pass (even with LessThan
and GreaterThan in the generators).  So things are looking
pretty rock solid.  The current approach creates 100
puzzles as part of the testing and makes sure that all
100 have unique solutions.