Interesting insights in studying difficulty
My original concept was a bit bogus.  I turned out that
it effectively amounted to just running all the primary
constraints over the sets of possible values and then
looking at what could be eliminated in that way.  But then
when I looked at the results, I noticed a pattern.  The tree
based approach tends to build up problems where the
first variable to be solved can be deduced strictly from
primary constraints (i.e. there is only one possible value
considering only primary constraint).  Then you tended
to see the next one having been reduced to 4 possible
values (basically, the same as the number of rows/columns).
This is because the primaries tend to reduce the number of
possibilities by a factor of 4.  Then the next one was 16
(another factor of 4).

The bottom line was that scoring just amounted to adding
up the number of possibilities that existed across all
variables after all primary constraints had been applied.